+++ tests/common.py
@@ -81,3 +81,13 @@ async def fake_post_request(*_, **kwargs):
 async def fake_post_request_multi(*args, **kwargs):
     kwargs["POSTFIX"] = "multi"
     return await fake_post_request(*args, **kwargs)
+
+
+async def fake_post_request_ac(*args, **kwargs):
+    kwargs["POSTFIX"] = "ac"
+    return await fake_post_request(*args, **kwargs)
+
+
+async def fake_post_request_bridged(*args, **kwargs):
+    kwargs["POSTFIX"] = "bridged"
+    return await fake_post_request(*args, **kwargs)